refactor: Document -> Dokument
RISDEV-7626
